Edited WinWall serializers, views based on Insomnia testing errors
In WinWallSerializer, sticky_id is not required so I commented it out (you actually don't need a sticky_id to create a win wall but you need a win wall to create a sticky note). Also imported the user.model SheCodesUser to work with winwall's models.

In WinWallList views, replaced "owner" by "user_id" as initially defined in the models.

WinWall.user_id" must be a "SheCodesUser" instance